Fix infinite redirects for slo requests.
When user clicks 'Logout' in the page, a request is made to /?slo, then
it redirects with a slo SAMLRequest to the configured slo endpoint for
the idp; then the idp also redirects back to /?slo, causing an infinite
redirection loop.

This uses previous commit to create a simple way to handle SAMLRequest,
SAMLResponse and user initiated logouts.