Allow user to configure which field used as attribute name
attribute names are currently identified by their OASIS spec
oid.  while these are specific, they're not easy to read.  Most
IdPs also include a "FriendlyName" field such as:

<saml2:Attribute FriendlyName="givenName"
    Name="urn:oid:2.5.4.42"
    NameFormat="urn:oasis:names:tc:SAML:2.0:attrname-format:uri">

That's a lot easier to use. Added support to advanced configuration
to allow the user to define the field name used. Defaults to Name if
not defined.